Pushed local project of live site to GitHub
This project was initially developed locally and then pushed to GitHub for version control and collaboration. By pushing the project to GitHub, the code is now accessible to other developers who can contribute to the project and track changes over time. This also allows for easy deployment of the live site from the repository.